Fix for unsafe column names in Spark SQL statements
The fix for ensuring that all column names that are used in Spark SQL
statements involves ensuring that all calls that would put the bare
column name into a Spark SQL statement do so only after calling
`ColumnName.sanitize`, which will wrap the name in backticks ("``").
Additionally, this fix includes understanding when it is crucial to
include the escaped column name vs. the real (unescaped) name.